+++ index.js
@@ -27,7 +27,7 @@ class Row extends React.Component {
     return false
   }
 
-  handleLongPress = e => {
+  handlePress = e => {
     if (!this.refs.view) return
     this.refs.view.measure(
       (frameX, frameY, frameWidth, frameHeight, pageX, pageY) => {
@@ -44,9 +44,10 @@ class Row extends React.Component {
   componentDidUpdate(props) {
     // Take a shallow copy of the active data. So we can do manual comparisons of rows if needed.
     if (props.rowHasChanged) {
-      this._data = typeof props.rowData.data === 'object'
-        ? Object.assign({}, props.rowData.data)
-        : props.rowData.data
+      this._data =
+        typeof props.rowData.data === 'object'
+          ? Object.assign({}, props.rowData.data)
+          : props.rowData.data
     }
   }
 
@@ -70,10 +71,12 @@ class Row extends React.Component {
       ),
       {
         sortHandlers: {
-          onLongPress: this.handleLongPress,
+          onLongPress: !this.props.moveOnPressIn ? this.handlePress : null,
+          onPressIn: this.props.moveOnPressIn ? this.handlePress : null,
           onPressOut: this.props.list.cancel,
         },
-        onLongPress: this.handleLongPress,
+        onLongPress: !this.props.moveOnPressIn ? this.handleLongPress : null,
+        onPressIn: this.props.moveOnPressIn ? this.handlePress : null,
         onPressOut: this.props.list.cancel,
       }
     )
@@ -208,9 +211,10 @@ class SortableListView extends React.Component {
         }
         const itemHeight = this.state.active.layout.frameHeight
         const fromIndex = this.order.indexOf(this.state.active.rowData.index)
-        let toIndex = this.state.hovering === false
-          ? fromIndex
-          : Number(this.state.hovering)
+        let toIndex =
+          this.state.hovering === false
+            ? fromIndex
+            : Number(this.state.hovering)
         const up = toIndex > fromIndex
         if (up) {
           toIndex--
@@ -452,7 +456,7 @@ class SortableListView extends React.Component {
 
   componentDidMount() {
     InteractionManager.runAfterInteractions(() => {
-      this.timer = setTimeout(this.measureWrapper, 0)
+      this.timer = setTimeout(() => this && this.measureWrapper(), 0)
     })
   }
 
@@ -477,9 +481,11 @@ class SortableListView extends React.Component {
     const scrollEnabled =
       !this.state.active && this.props.scrollEnabled !== false
 
+    const ListViewComponent = this.props.ListViewComponent || ListView
+
     return (
       <View ref="wrapper" style={{ flex: 1 }} collapsable={false}>
-        <ListView
+        <ListViewComponent
           enableEmptySections
           {...this.props}
           {...this.state.panResponder.panHandlers}